refactor: move TerminalGestureListenerCallback to interfaces package; cleanup
