+++ app/src/main/java/com/coderGtm/yantra/interfaces/TerminalGestureListenerCallback.kt
@@ -1,0 +1,9 @@
+package com.coderGtm.yantra.interfaces
+
+
+interface TerminalGestureListenerCallback {
+    fun onSingleTap()
+    fun onDoubleTap()
+    fun onSwipeRight()
+    fun onSwipeLeft()
+}
